3.27 로그인 회원가입 로그아웃 완료 (추후 오류가 발견 시 수정) 회원 탈퇴기능 구현예정

+++ src/main/java/com/example/be2ndproject/shopping_mall/repository/Payment/Payment.java
@@ -18,7 +18,7 @@ public class Payment {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "payment_id")
-    private Long paymentId;
+    private Integer paymentId;
 
     @Column(name = "total_price", nullable = false)
     private Integer totalPrice;
