+++ src/main/java/com/example/be2ndproject/shopping_mall/repository/Reservation/Reservations.java
@@ -19,7 +19,7 @@ public class Reservations {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "reservation_num")
-    private Long reservationNum;
+    private Integer reservationNum;
 
     @ManyToOne
     @JoinColumn(name = "user_id", referencedColumnName = "user_id")
